+++ Basil/include/Platform/OpenGL/OpenGLTexture2D.h
@@ -11,9 +11,11 @@ namespace Basil
 	{
 		public:
 			OpenGLTexture2D(const std::string& path);
+			OpenGLTexture2D(uint32_t width, uint32_t height);
 			~OpenGLTexture2D();
 			uint32_t getWidth() const override;
 			uint32_t getHeight() const override;
+			void setData(void* data, uint32_t size) override;
 			void bind(unsigned int slot = 0) const override;
 		private:
 			std::string path;