+++ Basil/include/Renderer/Renderer2D.h
@@ -17,7 +17,7 @@ namespace Basil
 			static void endScene();
 			static void drawQuad(const glm::vec2& position, const glm::vec2& size, const glm::vec4& color);
 			static void drawQuad(const glm::vec3& position, const glm::vec2& size, const glm::vec4& color);
-			static void drawQuad(const glm::vec2& position, const glm::vec2& size, const Shared<Texture2D>& texture);
-			static void drawQuad(const glm::vec3& position, const glm::vec2& size, const Shared<Texture2D>& texture);
+			static void drawQuad(const glm::vec2& position, const glm::vec2& size, float textureScale, const Shared<Texture2D>& texture);
+			static void drawQuad(const glm::vec3& position, const glm::vec2& size, float textureScale, const Shared<Texture2D>& texture);
 	};
 }