+++ Basil/include/Renderer/Texture.h
@@ -11,6 +11,7 @@ namespace Basil
 			virtual ~Texture() {}
 			virtual uint32_t getWidth() const = 0;
 			virtual uint32_t getHeight() const = 0;
+			virtual void setData(void* data, uint32_t size) = 0;
 			virtual void bind(unsigned int slot = 0) const = 0;
 	};
 }