+++ Basil/include/Renderer/Texture2D.h
@@ -14,5 +14,6 @@ namespace Basil
 	{
 		public:
 			static Shared<Texture2D> create(const std::string& path);
+			static Shared<Texture2D> create(uint32_t width, uint32_t height);
 	};
 }