+++ Basil/src/Platform/OpenGL/OpenGLTexture2D.cpp
@@ -55,6 +55,25 @@ namespace Basil
 		stbi_image_free(data);
 	}
 
+	OpenGLTexture2D::OpenGLTexture2D(uint32_t width, uint32_t height)
+	{
+		// Set variables
+		this->width = width;
+		this->height = height;
+		internalFormat = GL_RGBA8;
+		dataFormat = GL_RGBA;
+
+		// Create texture in OpenGL
+		glCreateTextures(GL_TEXTURE_2D, 1, &rendererID);
+		glTextureStorage2D(rendererID, 1, internalFormat, this->width, this->height);
+
+		// Set texture parameters
+		glTextureParameteri(rendererID, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
+		glTextureParameteri(rendererID, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
+		glTextureParameteri(rendererID, GL_TEXTURE_WRAP_S, GL_REPEAT);
+		glTextureParameteri(rendererID, GL_TEXTURE_WRAP_T, GL_REPEAT);
+	}
+
 	// Destructor
 	OpenGLTexture2D::~OpenGLTexture2D()
 	{
@@ -71,6 +90,14 @@ namespace Basil
 	uint32_t OpenGLTexture2D::getHeight() const
 	{
 		return height;
+	}
+
+	// Set texture data
+	void OpenGLTexture2D::setData(void* data, uint32_t size)
+	{
+		uint32_t bytesPerPixel = dataFormat == GL_RGBA ? 4 : 3;
+		ASSERT(size == width * height * bytesPerPixel, "Data must be entire texture!");
+		glTextureSubImage2D(rendererID, 0, 0, 0, width, height, dataFormat, GL_UNSIGNED_BYTE, data);
 	}
 
 	// Bind the texture to the provided slot