+++ Sandbox/assets/shaders/Texture.glsl
@@ -22,9 +22,11 @@ layout(location = 0) out vec4 color;
 
 in vec2 v_TexCoord;
 
+uniform float u_TexScale;
 uniform sampler2D u_Texture;
+uniform vec4 u_Color;
 
 void main()
 {
-	color = texture(u_Texture, v_TexCoord * 10.0);
+	color = texture(u_Texture, v_TexCoord * u_TexScale) * u_Color;
 }